+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [3.2.0] 2017-01-25 
+### Added
+- add `--self-check` option for gulp doctor which checks if `ecc-gulp-tasks` and `ecc-dotfiles` are up to date.
+
+### Changed
+- moved definitions for tools to json file
+
 ## [3.1.2] 2017-01-11 
 ### Fixed
 - upgraded `ecc-license-checker`, whoch fixes bugs in `gulp licenses-yaml2json`